refactor: move content from bin to src/cli
